llamaindex/full-pipeline-templates/solar-panels/requirements.txt
streamlit>=1.30.0
llama-cloud>=0.1.0
llama-cloud-services>=0.1.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
nest-asyncio>=1.5.0